requirements.txt
# Core libraries
python-dotenv
openai

# Optional / already used in your project
pandas==2.2.3
numpy==1.26.4
pypdf==5.1.0
pdfplumber==0.11.4
markdown==3.7
faiss-cpu==1.13.2
chromadb==0.5.23
ollama==0.4.7
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
gradio==5.12.0
tqdm==4.67.1
python-dotenv==1.0.1